+++ packages/client/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @vercel/client
 
+## 15.1.5
+
+### Patch Changes
+
+- Add yarn Plug'n'Play files to default Vercel ignore list ([#13140](https://github.com/vercel/vercel/pull/13140))
+
+- Updated dependencies [[`ac8efbbd20e6d006dfd050c452cf3ef28f7bb9a7`](https://github.com/vercel/vercel/commit/ac8efbbd20e6d006dfd050c452cf3ef28f7bb9a7)]:
+  - @vercel/build-utils@10.5.0
+
+## 15.1.4
+
+### Patch Changes
+
+- Updated dependencies [[`47e5335cadc62398600c456c09120582adb25c88`](https://github.com/vercel/vercel/commit/47e5335cadc62398600c456c09120582adb25c88)]:
+  - @vercel/build-utils@10.4.0
+
 ## 15.1.3
 
 ### Patch Changes