+++ packages/node/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @vercel/node
 
+## 5.1.13
+
+### Patch Changes
+
+- [cli] get maxDuration from function configuration ([#13141](https://github.com/vercel/vercel/pull/13141))
+
+- Updated dependencies [[`ac8efbbd20e6d006dfd050c452cf3ef28f7bb9a7`](https://github.com/vercel/vercel/commit/ac8efbbd20e6d006dfd050c452cf3ef28f7bb9a7)]:
+  - @vercel/build-utils@10.5.0
+
+## 5.1.12
+
+### Patch Changes
+
+- Updated dependencies [[`47e5335cadc62398600c456c09120582adb25c88`](https://github.com/vercel/vercel/commit/47e5335cadc62398600c456c09120582adb25c88)]:
+  - @vercel/build-utils@10.4.0
+
 ## 5.1.11
 
 ### Patch Changes