+++ packages/static-build/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/static-build
 
+## 2.7.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/gatsby-plugin-vercel-builder@2.0.79
+
+## 2.7.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/gatsby-plugin-vercel-builder@2.0.78
+
 ## 2.7.3
 
 ### Patch Changes